chore: Fixup issues from Django app rename
Wow, that is kind of annoying to do. Hopefully all of the hanging
issues have been caught.
