+++ terraso_backend/apps/graphql/schema/projects.py
@@ -19,7 +19,7 @@ from graphene import relay
 from graphene_django import DjangoObjectType
 
 from apps.graphql.exceptions import GraphQLNotAllowedException
-from apps.soilproj.models import Project, ProjectMembership, Site
+from apps.project_management.models import Project, ProjectMembership, Site
 
 from .commons import BaseMutation, BaseWriteMutation, TerrasoConnection
 from .sites import SiteNode
